backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend for the Full Stack Web App."
authors = ["Josh Wang <hello@wangjoshuah.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
Flask-Compress = "^1.12"
Flask-Cors = "^3.0.10"
flask-smorest = "^0.38.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
mypy = "^0.961"
johnnydep = "^1.13"
pre-commit = "^2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
